# perf: use stream for `getDeclDeps`