```
runtime: deno 1.41.0 (aarch64-apple-darwin)

file:///Users/nemo/repo/stackgraph/graph/decl_deps_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------
array           3.55 ms/iter         281.3     (2.2 ms … 10.47 ms) 4.09 ms 6.55 ms 10.47 ms
stream          2.52 ms/iter         397.2     (1.87 ms … 6.01 ms) 2.61 ms 5.62 ms 6.01 ms

summary
  stream
   1.41x faster than array
```

- around ~40% performance improvement
- no noticable differences in IRL performance (~500 files)
- bottleneck seems to come from [`ReferenceFindableNode.findReferences()`](https://ts-morph.com/navigation/finding-references#finding-referencing-nodes)